requirements.txt
asgiref==3.5.2
boto3==1.26.13
botocore==1.29.13
Django==4.1.3
django-crispy-forms==1.14.0
django-ses==3.2.2
django-storages==1.13.1
jmespath==1.0.1
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
